requirement.txt
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.18.1
webdriver-manager==4.0.1
lxml==5.2.1
nltk==3.8.1
pymupdf==1.23.26
fpdf2==2.7.7
sentence-transformers==2.6.1
chromadb==0.4.24
langchain
langchain-community
faiss-cpu
python-pptx==0.6.23
numpy==1.26.4
tqdm==4.66.2
markdown
